Add depcheck preset (#2119)
Add a preset for the depcheck library that is applied to all projects in scm-ui. Some minor dependency issues were discovered and resolved along the way. Some features of depcheck did not work correctly or were missing, which is why the configuration file is a bit more verbose than desired.