Fix handling of old commits as new ones in new branches
If a new branch is pushed in Git without new commits (the branch
is created on a commit that already exists in the SCM-Manager repository
and pushed without any further commit), all ancestors of the commit
the branch points to have been treated as if they were new by the
hook changeset provider. This led to severe errors like wrong
push logs (by the pushlog plugin) or re-evaluated commit messages
by the commit message checker plugin.

This fixes this wrong behaviour. If new commits are not provided by
the pack parser, no commit will be treated as a new one.