Remove unconditional force push
This removes the unconditional force mode used while pushing changes made by the modify or the merge command in git repositories. The force mode led to the "deletion" of other changes on the same branch made between the creation of the working copy and the push to the core repository inside SCM-Manager.

Committed-by: Konstantin Schaper <konstantin.schaper@cloudogu.com>