Fix copying source code in Firefox incorrectly adding extra line breaks
There is an open bug in Firefox which causes extra line breaks to appear when copying if the selection contains a `span` with the `user-select: none;` styles.
This workaround removes this style and displays the line number in a `::before` instead. This has the same effect of the user not being able to select it and it also doesn't show up in multi-line selections but prevents Firefox from interpreting it incorrectly.

Committed-by: Eduard Heimbuch <eduard.heimbuch@cloudogu.com>