Fix check for lfs files
This fixes the check for LFS files used in git imports
and mirror command.

First this checks for LFS files recursively in the tree
(in other words, before LFS files were found only in
the root directory).

Second, this fixes a bug in the used JGit lfs filter by
checking, whether the checked tree entry is a normal
file. This prevents missing object exceptions during
the check when the repository contains a submodule.

Committed-by: Thomas Zerr <thomas.zerr@cloudogu.com>
Co-authored-by: René Pfeuffer <rene.pfeuffer@cloudogu.com>