Fix order of items in keyboard iterator
The previous keyboard iterator did not take into account, if multiple items deregistered at once.
Therefore, the keyboard iterator ran into an iterator invalidation bug, because the assigned index of an element can run out of bounds.
Elements with an out of bounds assigned index, would not be able to deregister and therefore stay in the list of registered items.
Therefore messing up the iteration logic.