Fix misleading error log
When a commit could not be found, JGit throws a MissingObjectException.
Because this extends IOException, the according catch block is executed
with a log with level ERROR and the message, that the repository could not
be opened. This is simply wrong.

With this, we catch the MissingObjectException explicitly and simply return null.